+++ BeatPagination/BeatPagination/Pagination/BeatPaginator.m
@@ -315,7 +315,7 @@
 	else return BeatPaginator.lineHeight;
 }
 
-+ (NSInteger)heightForString:(NSString *)string font:(BeatFont *)font maxWidth:(NSInteger)maxWidth lineHeight:(CGFloat)lineHeight
++ (CGFloat)heightForString:(NSString *)string font:(BeatFont *)font maxWidth:(NSInteger)maxWidth lineHeight:(CGFloat)lineHeight
 {
 	/*
 	 This method MIGHT NOT work on iOS. For iOS you'll need to adjust the font size to 80% and use the NSString instance
